extend tests for getSearchUrl, change extension to .ts
